+++ src/main/java/chess/domain/ChessGame.java
@@ -1,5 +1,6 @@
 package chess.domain;
 
+import chess.domain.board.Board;
 import chess.domain.gamestate.Ready;
 import chess.domain.gamestate.State;
 import chess.domain.position.Position;
