+++ src/main/java/chess/domain/Score.java
@@ -1,5 +1,7 @@
 package chess.domain;
 
+import chess.domain.board.Board;
+
 public class Score {
     private final double blackScore;
     private final double whiteScore;
