+++ src/main/java/chess/domain/gamestate/Ready.java
@@ -1,6 +1,6 @@
 package chess.domain.gamestate;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.Score;
 import chess.domain.Side;
 import chess.domain.position.Position;
