+++ src/test/java/chess/domain/dao/BoardDAOTest.java
@@ -1,7 +1,7 @@
 package chess.domain.dao;
 
 import chess.dao.BoardDAO;
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.domain.Side;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.BeforeEach;
