+++ src/test/java/chess/domain/gamestate/GameSetTest.java
@@ -1,7 +1,7 @@
 package chess.domain.gamestate;
 
-import chess.domain.Board;
-import chess.domain.BoardInitializer;
+import chess.domain.board.Board;
+import chess.domain.board.BoardInitializer;
 import chess.domain.Side;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
