+++ src/test/java/chess/domain/gamestate/RunningTest.java
@@ -1,7 +1,7 @@
 package chess.domain.gamestate;
 
-import chess.domain.Board;
-import chess.domain.BoardTestInitializer;
+import chess.domain.board.Board;
+import chess.domain.board.BoardTestInitializer;
 import chess.domain.Side;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.BeforeEach;
